refactor: Update OpenFile and OpenS3 functions to use gzip compression
The OpenFile and OpenS3 functions have been updated to support gzip compression in the file and S3 providers. The code now creates a gzip writer with the best compression level when writing to files or S3 objects with a .gz or .gzip extension. If an error occurs while creating the gzip writer, an error message is logged and the function returns the error.

Refactor the code to improve gzip compression support in the file and S3 providers.